create util files for gutenberg functions and for language functions. pull first functions out of create to make the file more concise
